fix(spm): accumulate Unigram Viterbi scores in f64
f32 accumulation of partial path sums can flip exact score ties to a
different segmentation than HuggingFace tokenizers, which carries
backtrace_score in f64. Widen scores through the SentencePiece
tokenizer, its constructor, and the GGUF/HF-JSON loaders and Python
bindings that feed it, so tied lattice edges resolve identically to
the reference implementation.

+++ src/core/hf_json/loader.rs
@@ -234,7 +234,10 @@ fn build_unigram(root: &Value, model: &Value) -> Result<Backend, HfJsonError> {
             .first()
             .and_then(Value::as_str)
             .ok_or(HfJsonError::MissingField("model.vocab[*][0] = token"))?;
-        let score = pair.get(1).and_then(Value::as_f64).unwrap_or(0.0) as f32;
+        // Kept at full `f64` width: HF `tokenizers` reads the same JSON number
+        // into an `f64` and its Viterbi compares partial sums at that precision,
+        // so narrowing here would reorder equal-scoring segmentations.
+        let score = pair.get(1).and_then(Value::as_f64).unwrap_or(0.0);
         tokens.push(token.to_string());
         scores.push(score);
     }
